+++ Cargo.toml
@@ -7,7 +7,6 @@ edition = "2024"
 members = [
     "common",
     "ppu",
-    "ppu-v2",
     "cpu",
     "apu",
     "cpu/instr_metalang_procmacro",
