--- ppu-v2/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "ppu-v2"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-sdl2 = "0.36"
