+++ ppu/Cargo.toml
@@ -1,8 +1,7 @@
 [package]
-name = "ppu"
+name = "ppu-v2"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-minifb = "0.28"
-image = "0.25.6"
+sdl2 = "0.36"
