--- ppu/src/lib.rs
@@ -1,6 +1,0 @@
-pub mod utils;
-pub mod tile;
-pub mod ppu;
-
-#[cfg(test)]
-mod tests;
