--- ppu/src/tests/mod.rs
@@ -1,4 +1,0 @@
-mod ppu_cgram;
-mod ppu_vram;
-mod ppu_render;
-mod ppu_tile;
